+++ js/check_config.js
@@ -1,7 +1,7 @@
 const path = require("node:path");
 const fs = require("node:fs");
+const { styleText } = require("node:util");
 const Ajv = require("ajv");
-const colors = require("ansis");
 const globals = require("globals");
 const { Linter } = require("eslint");
 
@@ -35,7 +35,7 @@ function checkConfigFile () {
 
 	// Check permission
 	try {
-		fs.accessSync(configFileName, fs.F_OK);
+		fs.accessSync(configFileName, fs.constants.F_OK);
 	} catch (error) {
 		throw new Error(`${error}\nNo permission to access config file!`);
 	}
@@ -54,13 +54,14 @@ function checkConfigFile () {
 				globals: {
 					...globals.node
 				}
-			}
+			},
+			rules: { "no-undef": "error" }
 		},
 		configFileName
 	);
 
 	if (errors.length === 0) {
-		Log.info(colors.green("Your configuration file doesn't contain syntax errors :)"));
+		Log.info(styleText("green", "Your configuration file doesn't contain syntax errors :)"));
 		validateModulePositions(configFileName);
 	} else {
 		let errorMessage = "Your configuration file contains syntax errors :(";
@@ -72,6 +73,10 @@ function checkConfigFile () {
 	}
 }
 
+/**
+ *
+ * @param {string} configFileName - The path and filename of the configuration file to validate.
+ */
 function validateModulePositions (configFileName) {
 	Log.info("Checking modules structure configuration ...");
 
@@ -107,7 +112,7 @@ function validateModulePositions (configFileName) {
 
 	const valid = validate(data);
 	if (valid) {
-		Log.info(colors.green("Your modules structure configuration doesn't contain errors :)"));
+		Log.info(styleText("green", "Your modules structure configuration doesn't contain errors :)"));
 	} else {
 		const module = validate.errors[0].instancePath.split("/")[2];
 		const position = validate.errors[0].instancePath.split("/")[3];