+++ tests/e2e/vendor_spec.js
@@ -9,21 +9,13 @@ describe("Vendors", () => {
 	});
 
 	describe("Get list vendors", () => {
-		const vendors = require(`${__dirname}/../../vendor/vendor.js`);
+		const vendors = require(`${__dirname}/../../js/vendor.js`);
 
 		Object.keys(vendors).forEach((vendor) => {
 			it(`should return 200 HTTP code for vendor "${vendor}"`, async () => {
-				const urlVendor = `http://localhost:8080/vendor/${vendors[vendor]}`;
-				const res = await fetch(urlVendor);
-				expect(res.status).toBe(200);
-			});
-		});
-
-		Object.keys(vendors).forEach((vendor) => {
-			it(`should return 404 HTTP code for vendor https://localhost/"${vendor}"`, async () => {
 				const urlVendor = `http://localhost:8080/${vendors[vendor]}`;
 				const res = await fetch(urlVendor);
-				expect(res.status).toBe(404);
+				expect(res.status).toBe(200);
 			});
 		});
 	});